Render all sub-protocols belonging to a main protocol to single markdown files

The function should be called interactively (in the console) after the dependencies section in the YAML header of the index.Rmd file of the main protocol has been filled in with the aid of the protocolhelper::add_dependencies() function. For reproducibility, it is good practice to save the call in a separate R script. For each sub-protocol a single markdown file and associated media and data files will be written. Each sub-protocol will be written to a subfolder of the main protocol. The subfolder name is the same as the version number of the sub-protocol.

Usage

add_subprotocols(code_mainprotocol, fetch_remote = TRUE)

Arguments

code_mainprotocol

Character string giving the protocol code for the main protocol

fetch_remote

Whether or not to fetch the remote. Default TRUE.

See also