Helper function to check if author information is correct

Checks the format of author names and orcid ids.

check_all_author_info(author_list, problems_vect)

Arguments

author_list

the yaml front matter part containing author info

problems_vect

character vector of previously encountered problems

Value

a character vector of previously encountered problems and problems identified for author names and orcid ids.

See also