Create protocol code from it's components

A protocol code of format s[f|p]p-###-[nl|en] will be created. The number will be determined automatically based on theme (in case of sfp) and a rank order of all existing protocol numbers

create_protocol_code(protocol_type, theme, protocol_number, language)

Arguments

protocol_type
Either sfp (standard field protocol) or spp ( standard project protocol)

theme
A character string equal to one of "generic" (default), "water", "air", "soil", "vegetation" or "species". It is used as the folder location (src/thematic/theme) where standard field protocols that belong to the same theme will be stored. Ignored if protocol_type = "spp".

protocol_number
A character string giving the protocol number. This parameter should normally not be specified (i.e. NULL), unless from_docx is specified. A protocol number is a three digit string where the first digit corresponds with a theme and the last two digits identify a protocol within a theme for standard field protocols. A protocol number for a project-specific protocol is just a three digit string. If NULL (the default), a protocol number will be determined automatically based on pre-existing protocol numbers. Protocol numbers that are already in use can be retrieved with get_protocolnumbers().

language
Language of the protocol, either "nl" (Dutch), the default, or "en" (English).

Value

A character string containing the protocol_code