Surround chunks with pandoc markdown divs

Adapted from bookdown::fence_theorems() but with the following differences:

chunks will be kept, not replaced by divs

all chunks will be processed, not only theorem chunks

all chunks will be surrounded by div environments

a label will be added to the div environment based on the chunk name

if the chunk parameters contain fig.cap, the label will be #fig:<chunkname>, else we assume it is a table #tab:<chunkname>

fence_all_chunks(input)

Arguments

input

Path to input Rmarkdown file

Details

This function is only relevant in case of bookdown::markdown_document2 output in order to work around a problem with cross-references for tables and figures. It is assumed that the \@ref() syntax is used to refer to tables and figures and that chunk names are used as labels to refer to figures and tables. The input file will be overwritten.