Internal function to get (or set) the full path to a protocol. Either provide the full path of an existing, given folder name, or construct one for non-existing protocols using (new) folder name plus the theme or project name. For existing protocol folders, arguments theme and project_name are always ignored. To create a new protocol folder, also either the theme or the project_name argument are required apart from the protocol_folder_name.

Internal function to get (or set) the full path to a protocol.

Either provide the full path of an existing, given folder name, or construct one for non-existing protocols using (new) folder name plus the theme or project name.

For existing protocol folders, arguments theme and project_name are always ignored.

To create a new protocol folder, also either the theme or the project_name argument are required apart from the protocol_folder_name.

get_path_to_protocol(
  protocol_code,
  theme = NULL,
  project_name = NULL,
  short_title = NULL
)

Arguments

protocol_code
Character string giving the protocol code

theme
A character string equal to one of "generic", "water", "air", "soil", "vegetation" or "species". Defaults to NULL. Only needed if no folder with the name of the protocol code exists and the request is for a thematic protocol.

project_name
Character string giving the name of the project folder. Defaults to NULL. Only needed if no folder with the name of the protocol code exists and the request is for a project-specific protocol.

short_title
A character string of less than 20 characters to use in folder and filenames. Defaults to NULL. Only needed if no folder with the name of the protocol code exists.

Value

A character vector containing the full path to the protocol.

Examples

if (FALSE) {
get_path_to_protocol(protocol_code = "sfp-401-nl")
}