Fills in values from key-value pairs in yaml front matter

Overwrites the index.Rmd file

write_yaml_front_matter(
  parent_rmd,
  path_to_protocol,
  title,
  subtitle,
  date,
  reviewers,
  file_manager,
  version_number,
  protocol_code,
  language,
  authors,
  orcids,
  protocol_type,
  template,
  theme,
  project_name
)

Arguments

parent_rmd

original index.Rmd file

path_to_protocol

path to the protocol

title

A character string giving the main title of the protocol

subtitle

A character string for an optional subtitle

date

A character string of the date in ISO 8601 format (YYYY-MM-DD)

reviewers

A character vector for reviewers of the form First name Last name

file_manager

A character string for the name of the document maintainer of the form First name Last name

version_number

A version number of the form YYYY.##. The default is a function which will determine this number automatically. It should normally not be changed.

protocol_code

the protocol code

language

Language of the protocol, either "nl" (Dutch), the default, or "en" (English).

authors

A character vector for authors of the form c("lastname1, firstname1", "lastname2, firstname2")

orcids

A character vector of orcid IDs, equal in length to authors. If one of the authors does not have an orcid ID, use NA to indicate this in the corresponding position of the character vector (or get an orcid ID).

protocol_type

the protocol type

template

Which template to use? Default is set equal to protocol_type. However, you can also set this to "generic" in which case a simplified template will be used that can be used as an alternative to the default templates.

theme

A character string equal to one of "generic" (default), "water", "air", "soil", "vegetation" or "species". It is used as the folder location (source/sfp/theme) where standard field protocols that belong to the same theme will be stored. Ignored if protocol_type is other than "sfp".

project_name

A character string that is used as the folder location (source/spp/project_name) where project-specific protocols that belong to the same project will be stored. Preferably a short name or acronym. If the folder does not exist, it will be created. Ignored if protocol_type is other than "spp".