Create a folder with a bookdown (R markdown) template to start a new protocol and optionally render to html

This function will create a new folder based on values that are passed on via the parameters and creates a R-markdown (bookdown) skeleton based on a template file to start working on a new protocol. Optionally, the rmarkdown chapters are rendered to an html file which will be saved in a matching subfolder of the docs folder.

create_protocol(
  protocol_type = c("sfp", "spp"),
  title,
  subtitle,
  short_title,
  authors,
  orcids,
  date = Sys.Date(),
  reviewers,
  file_manager,
  version_number = paste0(format(Sys.Date(), "%Y"), ".00.dev"),
  theme = c("generic", "water", "air", "soil", "vegetation", "species"),
  project_name,
  language = c("nl", "en"),
  from_docx = NULL,
  protocol_number = NULL,
  render = FALSE
)

create_sfp(
  title,
  subtitle,
  short_title,
  authors,
  orcids,
  date = Sys.Date(),
  reviewers,
  file_manager,
  version_number = paste0(format(Sys.Date(), "%Y"), ".00.dev"),
  theme = c("generic", "water", "air", "soil", "vegetation", "species"),
  language = c("nl", "en"),
  from_docx = NULL,
  protocol_number = NULL,
  render = FALSE
)

create_spp(
  title,
  subtitle,
  short_title,
  authors,
  orcids,
  date = Sys.Date(),
  reviewers,
  file_manager,
  version_number = paste0(format(Sys.Date(), "%Y"), ".00.dev"),
  project_name,
  language = c("nl", "en"),
  from_docx = NULL,
  protocol_number = NULL,
  render = FALSE
)

Arguments

protocol_type
Either sfp (standard field protocol) or spp ( standard project protocol)

title
A character string giving the main title of the protocol

subtitle
A character string for an optional subtitle

short_title
A character string of less than 20 characters to use in folder and filenames

authors
A character vector for authors of the form First name Last name

orcids
A character vector of orcid IDs, equal in length to authors. If one of the authors does not have an orcid ID, use NA to indicate this in the corresponding position of the character vector (or get an orcid ID).

date
A character string of the date in ISO 8601 format (YYYY-MM-DD)

reviewers
A character vector for reviewers of the form First name Last name

file_manager
A character string for the name of the document maintainer of the form First name Last name

version_number
A version number of the form YYYY.##. For development versions .dev is added. The default is paste0(format(Sys.Date(), "%Y"), ".00.dev") (See from_docx). When the protocol is ready to be released, this should be changed by a repo- admin.

theme
A character string equal to one of "generic" (default), "water", "air", "soil", "vegetation" or "species". It is used as the folder location (src/thematic/theme) where standard field protocols that belong to the same theme will be stored. Ignored if protocol_type = "spp".

project_name
A character string that is used as the folder location (src/project/project_name) where project-specific protocols that belong to the same project will be stored. Preferably a short name or acronym. If the folder does not exist, it will be created. Ignored if protocol_type = "sfp".

language
Language of the protocol, either "nl" (Dutch), the default, or "en" (English).

from_docx
A character string with the path (absolute or relative) to a .docx file containing a pre-existing protocol. Please make sure to copy-paste all relevant meta-data from the .docx file to the corresponding parameters of this function. If nothing is provided (i.e. default = NULL), an empty template will be used.

protocol_number
A character string giving the protocol number. This parameter should normally not be specified (i.e. NULL), unless from_docx is specified. A protocol number is a three digit string where the first digit corresponds with a theme and the last two digits identify a protocol within a theme for standard field protocols. A protocol number for a project-specific protocol is just a three digit string. If NULL (the default), a protocol number will be determined automatically based on pre-existing protocol numbers. Protocol numbers that are already in use can be retrieved with get_protocolnumbers().

render
Whether or not to render the protocol to html. Defaults to FALSE.

Details

It is assumed that the src folder is a subfolder of an RStudio project with git version control. A target folder to which files will be written will be created as subdirectories beneath src. The subfolder structure is of the form /thematic/<theme>/<sfp>-<protocolnumber>-<language>_<short_title>/ for standard field protocols. Or /project/<project_name>/<spp>-<protocolnumber>-<language>_<short_title>/ for standard project protocols. The folder names are determined by the corresponding arguments of the function. A matching subfolder structure will be created beneath the docs folder (and output files needed for rendering to html output will be placed in it if render = TRUE. The template Rmarkdown files and the Rmarkdown files that result from converting a docx protocol (see from_docx argument), will be written to the target folder beneath src. Template Rmarkdown files with the same name as Rmarkdown files that result from converting a docx protocol will be overwritten by the latter. Besides Rmarkdown files, this target folder will also contain files needed to render to a Bookdown gitbook such as a _bookdown.yml. The NEWS.md file must be used to document the changes between revisions of the protocol. Furthermore, a data and a media folder will be created as subdirectories of the target folder. The media folder can be used to store image files and will contain image files extracted from the docx protocol when the from_docx argument is used. The data folder can be used to store tabular data that are needed for the protocol.